# Documentation/kernel/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cratis.Kernel.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/events/store/{microserviceId}/types": {
      "post": {
        "tags": [
          "EventTypes"
        ],
        "summary": "Register schemas.",
        "parameters": [
          {
            "name": "microserviceId",
            "in": "path",
            "description": "Cratis.Execution.MicroserviceId to register for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MicroserviceId"
            }
          }
        ],
        "requestBody": {
          "description": "The payload.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterEventTypes"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterEventTypes"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterEventTypes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "EventTypes"
        ],
        "summary": "Gets all event types.",
        "parameters": [
          {
            "name": "microserviceId",
            "in": "path",
            "description": "The Cratis.Execution.MicroserviceId to get event types for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MicroserviceId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventTypeInformation"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventTypeInformation"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventTypeInformation"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/events/store/{microserviceId}/types/schemas/{eventTypeId}": {
      "get": {
        "tags": [
          "EventTypes"
        ],
        "summary": "Gets generation schema for type.",
        "parameters": [
          {
            "name": "microserviceId",
            "in": "query",
            "description": "The Cratis.Execution.MicroserviceId to get event type for.",
            "schema": {
              "$ref": "#/components/schemas/MicroserviceId"
            }
          },
          {
            "name": "eventTypeId",
            "in": "path",
            "description": "Type to get for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EventTypeId"
            }
          },
          {
            "name": "microserviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/events/store/{microserviceId}/{tenantId}/observers": {
      "get": {
        "tags": [
          "Observers"
        ],
        "summary": "Get and observe all observers.",
        "parameters": [
          {
            "name": "microserviceId",
            "in": "path",
            "description": "Cratis.Execution.MicroserviceId the observers are for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MicroserviceId"
            }
          },
          {
            "name": "tenantId",
            "in": "path",
            "description": "Cratis.Execution.TenantId the observers are for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TenantId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ObserverState"
                    }
                  },
                  "description": "Represents an implementation of Cratis.Applications.Queries.IClientObservable."
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ObserverState"
                    }
                  },
                  "description": "Represents an implementation of Cratis.Applications.Queries.IClientObservable."
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ObserverState"
                    }
                  },
                  "description": "Represents an implementation of Cratis.Applications.Queries.IClientObservable."
                }
              }
            }
          }
        }
      }
    },
...
    "/api/events/store/{microserviceId}/projections": {
      "post": {
        "tags": [
          "Projections"
        ],
        "summary": "Register projections with pipelines.",
        "parameters": [
          {
            "name": "microserviceId",
            "in": "path",
            "description": "Cratis.Execution.MicroserviceId to register for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MicroserviceId"
            }
          }
        ],
        "requestBody": {
          "description": "The registrations.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterProjections"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterProjections"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterProjections"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Projections"
        ],
        "summary": "Gets all projections.",
        "parameters": [
          {
            "name": "microserviceId",
            "in": "path",
            "description": "The Cratis.Execution.MicroserviceId to get projections for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MicroserviceId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Projection"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Projection"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Projection"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/events/store/{microserviceId}/projections/immediate/{tenantId}": {
      "post": {
        "tags": [
          "Projections"
        ],
        "summary": "Perform an immediate projection.",
        "parameters": [
          {
            "name": "microserviceId",
            "in": "path",
            "description": "Cratis.Execution.MicroserviceId to perform for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MicroserviceId"
            }
          },
          {
            "name": "tenantId",
            "in": "path",
            "description": "Cratis.Execution.TenantId to perform for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TenantId"
            }
          }
        ],
        "requestBody": {
          "description": "The details about the Cratis.API.Projections.Commands.ImmediateProjection.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImmediateProjection"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ImmediateProjection"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ImmediateProjection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImmediateProjectionResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImmediateProjectionResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImmediateProjectionResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventSequenceId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "uuid"
          },
          "isEventLog": {
            "type": "boolean",
            "description": "Get whether or not this is the default log event sequence.",
            "readOnly": true
          },
          "isOutbox": {
            "type": "boolean",
            "description": "Get whether or not this is the default outbox event sequence.",
            "readOnly": true
          },
          "isInbox": {
            "type": "boolean",
            "description": "Get whether or not this is the default outbox event sequence.",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Represents the unique identifier of an event sequence."
      },
...
      "EventSequenceNumber": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "int64"
          }
        },
        "additionalProperties": false,
        "description": "Represents the sequence number within an event log for an event."
      },
...
      "EventType": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/EventTypeId"
          },
          "generation": {
            "$ref": "#/components/schemas/EventGeneration"
          },
          "isPublic": {
            "type": "boolean",
            "description": "Whether or not the event type is considered a public event."
          }
        },
        "additionalProperties": false,
        "description": "Represents the type of an event."
      },
      "EventTypeId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Represents the concept of the unique identifier of a type of event."
      },
      "EventTypeInformation": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "generations": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "EventTypeRegistration": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/EventType"
          },
          "friendlyName": {
            "type": "string",
            "description": "A friendly name.",
            "nullable": true
          },
          "schema": {
            "$ref": "#/components/schemas/JsonNode"
          }
        },
        "additionalProperties": false,
        "description": "Representation of an event type registration."
      },
      "FailedObserverPartition": {
        "type": "object",
        "properties": {
          "eventSourceId": {
            "$ref": "#/components/schemas/EventSourceId"
          },
          "sequenceNumber": {
            "$ref": "#/components/schemas/EventSequenceNumber"
          },
          "occurred": {
            "type": "string",
            "description": "Gets or sets the occurred time of the failure - if any.",
            "format": "date-time"
          },
          "lastAttempt": {
            "type": "string",
            "description": "Gets or sets the occurred time of the failure - if any.",
            "format": "date-time"
          },
          "attempts": {
            "type": "integer",
            "description": "Gets or sets the number of retry attempts it has had.",
            "format": "int32"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets or sets the message from the failure - if any.",
            "nullable": true
          },
          "stackTrace": {
            "type": "string",
            "description": "Gets or sets the stack trace from the failure - if any.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Represents the state used for failed observers."
      },
...
      "ImmediateProjection": {
        "type": "object",
        "properties": {
          "projectionId": {
            "$ref": "#/components/schemas/ProjectionId"
          },
          "eventSequenceId": {
            "$ref": "#/components/schemas/EventSequenceId"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelKey"
          },
          "projection": {
            "$ref": "#/components/schemas/JsonNode"
          }
        },
        "additionalProperties": false,
        "description": "Represents the payload for performing an immediate projection."
      },
      "ImmediateProjectionResult": {
        "type": "object",
        "properties": {
          "model": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonNode"
            },
            "description": "The Json representation of the model.",
            "nullable": true
          },
          "affectedProperties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyPath"
            },
            "description": "Collection of properties that was set.",
            "nullable": true
          },
          "projectedEventsCount": {
            "type": "integer",
            "description": "Number of events that caused projection.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Represents the result of an Cratis.Kernel.Grains.Projections.IImmediateProjection."
      },
...
      "MicroserviceId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Represents the concept of the microservice identifier."
      },
...
      "ObserverId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Concept that represents the unique identifier of an observer."
      },
      "ObserverName": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Concept that represents the name of an observer."
      },
      "ObserverRunningState": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11
        ],
        "type": "integer",
        "description": "Defines the status of an observer.",
        "format": "int32"
      },
      "ObserverState": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the identifier of the observer state.",
            "nullable": true
          },
          "eventTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventType"
            },
            "description": "Gets or sets the event types the observer is observing.",
            "nullable": true
          },
          "eventSequenceId": {
            "$ref": "#/components/schemas/EventSequenceId"
          },
          "observerId": {
            "$ref": "#/components/schemas/ObserverId"
          },
          "name": {
            "$ref": "#/components/schemas/ObserverName"
          },
          "type": {
            "$ref": "#/components/schemas/ObserverType"
          },
          "nextEventSequenceNumber": {
            "$ref": "#/components/schemas/EventSequenceNumber"
          },
          "lastHandled": {
            "$ref": "#/components/schemas/EventSequenceNumber"
          },
          "runningState": {
            "$ref": "#/components/schemas/ObserverRunningState"
          },
          "failedPartitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FailedObserverPartition"
            },
            "description": "Gets or sets the failed partitions for the observer.",
            "nullable": true
          },
          "recoveringPartitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecoveringFailedObserverPartition"
            },
            "description": "Gets or sets the failed partitions for the observer.",
            "nullable": true
          },
          "hasFailedPartitions": {
            "type": "boolean",
            "description": "Gets whether or not there are any failed partitions.",
            "readOnly": true
          },
          "isRecoveringAnyPartition": {
            "type": "boolean",
            "description": "Gets whether or not there are any partitions being recovered.",
            "readOnly": true
          },
          "isDisconnected": {
            "type": "boolean",
            "description": "Gets whether or not the observer is in disconnected state. Meaning that there is no subscriber to it.",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Represents the state used for an observer."
      },
      "ObserverType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "description": "Defines the different types of observers.",
        "format": "int32"
      },
...
      "Projection": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ProjectionId"
          },
          "name": {
            "$ref": "#/components/schemas/ProjectionName"
          },
          "modelName": {
            "$ref": "#/components/schemas/ModelName"
          }
        },
        "additionalProperties": false
      },
...
      "ProjectionRegistration": {
        "type": "object",
        "properties": {
          "projection": {
            "$ref": "#/components/schemas/JsonNode"
          },
          "pipeline": {
            "$ref": "#/components/schemas/JsonNode"
          }
        },
        "additionalProperties": false,
        "description": "Represents a single projection registration."
      },
...
      "RegisterEventTypes": {
        "type": "object",
        "properties": {
          "types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventTypeRegistration"
            },
            "description": "Collection of Cratis.API.Projections.Commands.EventTypeRegistration.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Payload for registering multiple event types."
      },
      "RegisterProjections": {
        "type": "object",
        "properties": {
          "projections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectionRegistration"
            },
            "description": "Collection of Cratis.API.Projections.Commands.ProjectionRegistration.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Represents the payload for registering projections."
      },
...
      "TenantId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Represents the unique identifier of a tenant in the system."
      },
...
    }
  }
}